[x86][PATCH] Fix Scheduling while atomic issue with kernel_write
On SMP systems the kernel seems to yield "Scheduling while atomic" bug,
with a complain that preemption was disabled somewhere at init_module.
preempt_enable seems to fix it, though since I'm not an expert with
preemption and scheduling I'm not sure this is the right way to fix that issue.

Lastly, the module build system outputs another file "lime.mod", then
"*.mod" is added to .gitignore.

ENV:
6.0.0-12parrot1-amd64
LiME disk format build
Built and loaded inside a virtual machine

Signed-off-by: Roi L <roeilev321_@outlook.com>